Implement collect_tokens function
There is a bug where given an "+()" instead of the expected
-> [Plus; LParen; RParen] it gives -> [Plus; Plust; LParen; RParen]
in other words it repeats whatever the first token it recognizes is.
I suspect that is a bug in either the collect_tokens function or the
advance function that advances the lexer. An easy fix would be to take
off the head of the list of the collect_tokens function after it's been
computed. However, seems like a bandaid solution. Will investigate this
later.